+++ src/main/java/io/vertx/ext/asyncsql/impl/ScalaUtils.java
@@ -16,21 +16,31 @@
 
 package io.vertx.ext.asyncsql.impl;
 
+import com.github.mauricio.async.db.RowData;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
+import io.vertx.core.json.JsonArray;
+import org.joda.time.DateTime;
+import org.joda.time.LocalDate;
+import org.joda.time.LocalDateTime;
 import scala.Function1;
 import scala.collection.immutable.List;
 import scala.concurrent.ExecutionContext;
 import scala.runtime.AbstractFunction1;
 import scala.util.Try;
 
+import java.time.Instant;
+import java.util.UUID;
+
 /**
  * Some Scala <=> Java conversion utilities.
  *
  * @author <a href="http://escoffier.me">Clement Escoffier</a>
  */
-public class ScalaUtils {
+public final class ScalaUtils {
+
+  private ScalaUtils () {}
 
   public static <T> Future<T> scalaToVertx(scala.concurrent.Future<T> future, ExecutionContext ec) {
     Future<T> fut = Future.future();
@@ -85,5 +95,31 @@ public class ScalaUtils {
         return null;
       }
     };
+  }
+
+  public static JsonArray rowToJsonArray(RowData data) {
+    JsonArray array = new JsonArray();
+    data.foreach(new AbstractFunction1<Object, Void>() {
+      @Override
+      public Void apply(Object value) {
+        if (value == null) {
+          array.addNull();
+        } else if (value instanceof scala.math.BigDecimal) {
+          array.add(value.toString());
+        } else if (value instanceof LocalDateTime) {
+          array.add(value.toString());
+        } else if (value instanceof LocalDate) {
+          array.add(value.toString());
+        } else if (value instanceof DateTime) {
+          array.add(Instant.ofEpochMilli(((DateTime) value).getMillis()));
+        } else if (value instanceof UUID) {
+          array.add(value.toString());
+        } else {
+          array.add(value);
+        }
+        return null;
+      }
+    });
+    return array;
   }
 }
